requirements.txt
# Core dependencies
flask==3.0.0                      # Flask for API
gunicorn==21.2.0                   # Production WSGI server

# Machine Learning & Transformers
torch==2.1.0                       # PyTorch (Ensure compatibility with CUDA)
transformers==4.39.0                # Hugging Face Transformers
sentence-transformers==2.2.2        # Sentence Transformers for MiniLM
accelerate==0.26.0                  # Optimization for large models like LLAMA

# FAISS for Vector Search
faiss-cpu==1.7.4                    # FAISS (CPU version)
# If using GPU, replace above with:
# faiss-gpu==1.7.4                  # FAISS (GPU version)

# NumPy & JSON Handling
numpy==1.26.0                       # For numerical computations
jsonschema==4.19.0                   # JSON schema validation

# FastAPI (Optional, if replacing Flask)
fastapi==0.110.0                    # Alternative to Flask for better scalability
uvicorn==0.29.0                      # ASGI server for FastAPI

# Web Framework & Frontend
jinja2==3.1.2                        # HTML templating for Flask UI
d3js==7.8.5                          # D3.js for visualizations (Optional)

# CUDA Dependencies (if running on GPU)
# pip install nvidia-pyindex
# pip install nvidia-cuda-runtime-cu12
